cleaned up base class
